producer/requirements.txt
redis==2.10.6
kafka-python==2.0.1
requests==2.28.2
pandas==2.2.2
python-dotenv==1.0.1


